Allow to reuse the flasher stub
1. "--after no_reset_stub" has been added which will keep the flasher
   stub running.
2. The ESP_SYNC command is implemented in the flasher stub. It will give
   similar response to the bootloader.

This helps in the following case. The chip cannot be properly reseted
and will remain in flasher stub. The next run of esptool won't be able
to reset the chip into bootloader and will talk to the flasher stub. A
proper response to ESP_SYNC is expected to establish connection.

Closes https://github.com/espressif/esptool/issues/603